paso initUI a HomeFragment
